+++ src/service/worker/runtime/builtIn/middlewares.ts
@@ -1,42 +1,13 @@
-import Koa from 'koa'
 import { collectDefaultMetrics, Gauge, Registry } from 'prom-client'
 import gcStats from 'prometheus-gc-stats'
 
-import { MetricsLogger } from '../../logger/metricsLogger'
-import { ServiceContext, ServiceJSON, ServiceRuntimeContext } from './typings'
-import { createRecorder } from './utils/recorder'
+import { MetricsLogger } from '../../../logger/metricsLogger'
+import { ServiceContext, ServiceRuntimeContext } from '../typings'
+import { createRecorder } from '../utils/recorder'
 
-export function recorderMiddleware (ctx: Koa.Context, next: () => Promise<void>) {
+export function recorderMiddleware (ctx: ServiceContext, next: () => Promise<void>) {
   ctx.state.recorder = createRecorder(ctx)
   return next()
-}
-
-export const whoAmIHandler = ({
-  events,
-  routes,
-}: ServiceJSON) => (ctx: ServiceContext) => {
-  ctx.status = 200
-  ctx.body = {
-    events,
-    routes,
-  }
-  ctx.set('Cache-Control', 'public, max-age=86400') // cache for 24 hours
-}
-
-export const healthcheckHandler = ({
-  events,
-  routes,
-}: ServiceJSON) => (ctx: ServiceContext) => {
-  ctx.status = 200
-  ctx.body = {
-    events,
-    routes,
-  }
-}
-
-export const metricsLoggerHandler = (ctx: ServiceRuntimeContext) => {
-  ctx.status = 200
-  ctx.body = ctx.metricsLogger.getSummaries()
 }
 
 export const addMetricsLoggerMiddleware = () => {
